Move summons spawn handling to separate module
